+++ demo/node/antd-pro/testv3/src/pages/component/_layout.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import router from 'umi/router';
 
 export default (props: any) => {
-  console.log('props => ', props);
+  console.log('props.route => ', props.route);
   if (props.location.pathname === '/compontent') {
     router.push('/compontent/list');
     return <div />;
